+++ frontend_v2/src/app/components/challenge/challenge-routing.module.ts
@@ -12,7 +12,6 @@ import { ChallengesubmissionsComponent } from './challengesubmissions/challenges
 import { ChallengeviewallsubmissionsComponent } from './challengeviewallsubmissions/challengeviewallsubmissions.component';
 import { ChallengeleaderboardComponent } from './challengeleaderboard/challengeleaderboard.component';
 import { ChallengesettingsComponent } from './challengesettings/challengesettings.component';
-import { ChallengemanageComponent } from './challengemanage/challengemanage.component';
 import { ChallengediscussComponent } from './challengediscuss/challengediscuss.component';
 
 const routes: Routes = [
@@ -34,7 +33,6 @@ const routes: Routes = [
       { path: 'leaderboard/:split', component: ChallengeleaderboardComponent },
       { path: 'leaderboard/:split/:entry', component: ChallengeleaderboardComponent },
       { path: 'settings', component: ChallengesettingsComponent },
-      { path: 'manage', component: ChallengemanageComponent },
       { path: 'discuss', component: ChallengediscussComponent },
     ],
   },
