+++ frontend_v2/src/app/components/challenge/challenge.module.ts
@@ -30,7 +30,6 @@ import { ChallengeComponent } from './challenge.component';
 import { ChallengeRoutingModule } from './challenge-routing.module';
 import { SharedModule } from '../../shared/shared.module';
 import { ChallengelistModule } from '../publiclists/challengelist/challengelist.module';
-import { ChallengemanageComponent } from './challengemanage/challengemanage.component';
 import { SubmissionMetaAttributesDialogueComponent } from './submission-meta-attributes-dialogue/submission-meta-attributes-dialogue.component';
 import { ChallengediscussComponent } from './challengediscuss/challengediscuss.component';
 
@@ -47,7 +46,6 @@ import { ChallengediscussComponent } from './challengediscuss/challengediscuss.c
     ChallengesubmissionsComponent,
     ChallengeviewallsubmissionsComponent,
     PhasecardComponent,
-    ChallengemanageComponent,
     SubmissionMetaAttributesDialogueComponent,
     ChallengediscussComponent,
   ],
