+++ frontend_v2/src/app/components/challenge/challengesettings/challengesettings.component.html
@@ -99,5 +99,42 @@
         </div>
       </div>
     </div>
+    <div class="row settings-section">
+      <div class="row margin-bottom-cancel">
+        <div class="col s12">
+          <h5 class="w-300">Submission worker actions</h5>
+        </div>
+      </div>
+      <div class="row margin-bottom-cancel">
+        <div class="col s12">
+          <button
+            class="btn ev-btn-dark waves-effect waves-dark grad-btn grad-btn-dark fs-12"
+            (click)="manageWorker('start')"
+          >
+            <i class="fa fa-play"></i> Start worker
+          </button>
+          <button
+            class="btn ev-btn-dark waves-effect waves-dark grad-btn grad-btn-dark fs-12"
+            (click)="manageWorker('stop')"
+          >
+            <i class="fa fa-stop"></i> Stop worker
+          </button>
+          <button
+            class="btn ev-btn-dark waves-effect waves-dark grad-btn grad-btn-dark fs-12"
+            (click)="manageWorker('restart')"
+          >
+            <i class="fa fa-refresh"></i> Restart worker
+          </button>
+        </div>
+      </div>
+      <div class="row margin-bottom-cancel">
+        <div class="col s12">
+          <h5 class="w-300">Submission worker logs</h5>
+        </div>
+      </div>
+      <div class="ev-logs-card">
+        <div *ngFor="let log of workerLogs; index as i" class="ev-logs">{{ log }}</div>
+      </div>
+    </div>
   </div>
 </div>
